Migrate fjs/ci/nix to .f.mjs, split NixJob type into types.ts
Updates the two dependents (fjs/ci, fjs/ci/node) to import runtime
values from .f.mjs and types from types.ts separately, and fixes
stale module.f.ts doc references including two with stale
line-number citations.
