+++ fjs/ci/module.f.ts
@@ -24,7 +24,8 @@ import type {
 } from './common/types.ts'
 import { rustPlatformSteps, rustWasmSteps } from './rust/module.f.mjs'
 import { nodeMainSteps, nodeNixJobs, nodeNixVersionSteps, nodeVersionJobs } from './node/module.f.ts'
-import { nixFlakes, nixInstall, type NixJob } from './nix/module.f.ts'
+import { nixFlakes, nixInstall } from './nix/module.f.mjs'
+import type { NixJob } from './nix/types.ts'
 import { bunSteps } from './bun/module.f.mjs'
 import { denoSteps } from './deno/module.f.mjs'
